requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==26.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-dotenv==1.0.1
sqlparse==0.5.1
whitenoise==6.7.0
